Organizational commit: moved .css files to new folder called "stylesheets"
