+++ snackpacks-website/src/containers/BlackList.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { ListGroup, ListGroupItem } from "react-bootstrap";
 import { LinkContainer } from "react-router-bootstrap";
-import "./DriverList.css";
+import "./stylesheets/BlackList.css";
 
 export default class DriverList extends Component {
     constructor(props) {
