+++ snackpacks-website/src/containers/NotFound.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./NotFound.css";
+import "./stylesheets/NotFound.css";
 
 export default () =>
     <div className="NotFound">
