+++ snackpacks-website/src/containers/RefundRequest.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import LoaderButton from "../components/LoaderButton";
-import "./Drivers.css";
+import "./stylesheets/RefundRequest.css";
 import {ListGroup, ListGroupItem} from "react-bootstrap";
 
 export default class SnackPacks extends Component {
@@ -119,7 +119,7 @@ export default class SnackPacks extends Component {
 
     render() {
         return (
-            <div className="Drivers">
+            <div className="RefundRequest">
                 <ListGroup>
                     <br></br>
                     <h3>{"Refund Request #" + (this.state.number+1) + ":"}</h3>
