+++ snackpacks-website/src/containers/UserList.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { ListGroup, ListGroupItem } from "react-bootstrap";
 import { LinkContainer } from "react-router-bootstrap";
-import "./DriverList.css";
+import "./stylesheets/UserList.css";
 
 export default class DriverList extends Component {
     constructor(props) {
