+++ snackpacks-website/src/containers/stylesheets/BlackListUsers.css
@@ -1,0 +1,8 @@
+.BlackListUsers form BlackListUsers{
+    padding-bottom: 30px;
+}
+
+.BlackListUsers form textarea {
+    height: 20px;
+    font-size: 16px;
+}
