+++ snackpacks-website/src/containers/stylesheets/DriverList.css
@@ -1,0 +1,12 @@
+
+.DriverList h4 {
+    font-family: "Open Sans", sans-serif;
+    font-weight: 200;
+    overflow: hidden;
+    line-height: 1.5;
+    white-space: nowrap;
+    text-overflow: ellipsis;
+}
+.DriverList p {
+    color: #666;
+}
