+++ snackpacks-website/src/containers/stylesheets/RefundRequest.css
@@ -1,0 +1,8 @@
+.RefundRequest form RefundRequest{
+    padding-bottom: 30px;
+}
+
+.RefundRequest form textarea {
+    height: 20px;
+    font-size: 16px;
+}
