+++ snackpacks-website/src/containers/stylesheets/Users.css
@@ -1,0 +1,8 @@
+.Users form Users{
+    padding-bottom: 30px;
+}
+
+.Users form textarea {
+    height: 20px;
+    font-size: 16px;
+}
